feat: mid-range multipliers and long wallet credits
Add 10x/20x/50x for reward and damage. Store spendable credits as
long so high multipliers no longer wrap to negatives (v0.13.4).

+++ app/src/main/java/ch/logixisland/anuto/business/game/GameState.java
@@ -65,7 +65,7 @@ public class GameState implements ScoreBoard.Listener, Persister {
     }
 
     @Override
-    public void creditsChanged(int credits) {
+    public void creditsChanged(long credits) {
 
     }
