+++ app/src/main/java/ch/logixisland/anuto/business/tower/TowerSelector.java
@@ -172,7 +172,7 @@ public class TowerSelector implements ScoreBoard.Listener, Entity.Listener, Towe
     }
 
     @Override
-    public void creditsChanged(int credits) {
+    public void creditsChanged(long credits) {
         if (mTowerInfo != null) {
             updateTowerInfo();
         }
